requirements.txt
spotipy
yt-dlp
mutagen
python-dotenv
requests
tqdm
colorama
beautifulsoup4
pillow
pydub 